Change breakpoints to match bootstrap's defaults
Doing this as short-term solution b/c we plan on using the Land Of Coder
Mega Menu module in a lot of future projects. That module (and other LoC modules)
introduces and depends on bootstrap.

Long-term plan would be to develop a separate module to help override
the Land of Coder module(s) dependencies on bootstrap and replace it
with either a more agnostic SASS or biased towards Foundation 6.x